stories user and form valdation complete

+++ app/api/stories_routes.py
@@ -1,8 +1,19 @@
 from flask import Blueprint, request
 from app.models import Story, db, Comment, User
 from flask_login import login_required, current_user
+from app.forms import StoryForm
 
 stories_routes = Blueprint('stories', __name__)
+
+def validation_errors_to_error_messages(validation_errors):
+    """
+    Simple function that turns the WTForms validation errors into a simple list
+    """
+    errorMessages = []
+    for field in validation_errors:
+        for error in validation_errors[field]:
+            errorMessages.append(f'{field} : {error}')
+    return errorMessages
 
 @stories_routes.route('')
 def get_stories():
@@ -14,34 +25,44 @@ def get_stories():
 @stories_routes.route('', methods=['POST'])
 @login_required
 def post_story():
-    data = request.json
-    print(data)
-    story = Story(title=data['title'],
-                body=data['body'],
-                user_id=data['user_id']
-                )
-    db.session.add(story)
-    db.session.commit()
-    return story.to_dict()
+    form = StoryForm()
+    form['csrf_token'].data = request.cookies['csrf_token']
+    if form.validate_on_submit():
+        story = Story(title=form.data['title'],
+                    body=form.data['body'],
+                    user_id=current_user.id
+                    )
+        db.session.add(story)
+        db.session.commit()
+        return story.to_dict()
+    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
 
 @stories_routes.route('/<int:id>',  methods=['PUT'])
 @login_required
 def edit_story(id):
-    data = request.json
     story = Story.query.get(id)
-    story.title = data['title']
-    story.body = data['body']
-    db.session.add(story)
-    db.session.commit()
-    return story.to_dict()
+    if current_user.id == story.user_id:
+        form = StoryForm()
+        form['csrf_token'].data = request.cookies['csrf_token']
+        if form.validate_on_submit():
+            story.title = form.data['title']
+            story.body = form.data['body']
+            db.session.add(story)
+            db.session.commit()
+            return story.to_dict()
+        return {'errors': validation_errors_to_error_messages(form.errors)}, 401
+    return {'errors': ['Unauthorized']}
+
 
 @stories_routes.route('/<int:id>', methods=['DELETE'])
 @login_required
 def delete_story(id):
     story = Story.query.get(id)
-    db.session.delete(story)
-    db.session.commit()
-    return story.to_dict()
+    if current_user.id == story.user_id:
+        db.session.delete(story)
+        db.session.commit()
+        return story.to_dict()
+    return {'errors': ['Unauthorized']}
 
 @stories_routes.route('/<int:id>/comments')
 @login_required
@@ -68,7 +89,7 @@ def post_comment():
 
 # ====================likes stories====================================
 
-#get all likes 
+#get all likes
 @stories_routes.route('/')
 def get_likes_story():
     likes = like_story.query.all()
@@ -94,13 +115,13 @@ def post_like(id):
 
     # allStory = user.liked
     # print ("one story???", allStory[0].to_dict())
-    
+
     # # num = like_story.query.filter(story_id == id)
 
     # print("current user",current_user.id)
     # print ("all like story?????",allStory)
-  
-  
+
+
     return "like"
 
 
@@ -113,4 +134,3 @@ def delete_like(id):
     user.liked.remove(user)
     db.session.commit()
     return "unlike"
-
